requirements.txt
asgiref==3.3.1
Django==3.1.6
django-extensions==3.1.0
djangorestframework==3.12.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
pdoc3==0.9.2
psycopg2-binary==2.8.6
pytz==2020.5
PyYAML==5.4.1
sqlparse==0.4.1
